Refactor: Moved verbose decorator to helper
