+++ modules/helpers/helper.py
@@ -1,6 +1,8 @@
 import sys
 from io import StringIO
 
+import matplotlib.pyplot as plt
+
 
 class Capturing(list):
     """Capture stdout into a single buffer list.
@@ -22,3 +24,16 @@ class Capturing(list):
         self.extend(self._stringio.getvalue().splitlines())
         del self._stringio  # free up some memory
         sys.stdout = self._stdout
+
+
+def verbose(func):
+    """Verbose decorator"""
+
+    def wrapper(*args, **kwargs):
+        if args[0].verbose:
+            print(f"## {func.__doc__}...")
+        plt.cla()
+        plt.grid()
+        return func(*args, **kwargs)
+
+    return wrapper
